# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AuthCredentials {
  /**
     * @minLength 2
     * @maxLength 30
     */
  robloxUsername: string;
  /** @minLength 6 */
  password: string;
}

export interface UserRecord {
  id: number;
  robloxUsername: string;
  status: string;
  createdAt?: string;
}

export interface UserUpdate {
  /** @maxLength 60 */
  status?: string;
}

export interface AuthResponse {
  token: string;
  user: UserRecord;
}

export interface ProfileRecord {
  id: number;
  /** @nullable */
  userId?: number | null;
  username: string;
  displayName: string;
  favoriteColor: string;
  bio: string;
  imageData: string;
  ageGroup: string;
  badges: string[];
  traits: string[];
  skin: string;
  /** @nullable */
  banner?: string | null;
  createdAt: string;
}

export interface ProfileInput {
  /**
     * @minLength 1
     * @maxLength 30
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 30
     */
  displayName: string;
  favoriteColor: string;
  /** @maxLength 75 */
  bio: string;
  imageData: string;
  ageGroup: string;
  banner?: string;
  traits?: string[];
}

export interface ProfileUpdate {
  skin?: string;
  badges?: string[];
  /** @nullable */
  banner?: string | null;
  username?: string;
  displayName?: string;
  favoriteColor?: string;
  bio?: string;
  imageData?: string;
  ageGroup?: string;
  traits?: string[];
}

export interface EventRecord {
  id: number;
  title: string;
  description: string;
  imageData: string;
  createdAt: string;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description: string;
  imageData?: string;
}

export interface BulletinRecord {
  id: number;
  category: string;
  title: string;
  content: string;
  url: string;
  sortOrder: string;
  createdAt: string;
}

export interface BulletinInput {
  category: string;
  title: string;
  content?: string;
  url?: string;
  sortOrder?: string;
}

export interface CustomBannerRecord {
  id: number;
  name: string;
  patternType: string;
  primaryColor: string;
  secondaryColor: string;
  bgColor: string;
  createdAt: string;
}

export interface CustomBannerInput {
  /** @minLength 1 */
  name: string;
  patternType: string;
  primaryColor: string;
  secondaryColor: string;
  bgColor: string;
}

export interface CustomBadgeRecord {
  id: number;
  name: string;
  color: string;
  accessory: string;
  createdAt: string;
}

export interface CustomBadgeInput {
  /** @minLength 1 */
  name: string;
  color: string;
  accessory: string;
}

export interface CustomSkinRecord {
  id: number;
  name: string;
  glowColor: string;
  glowEnabled: boolean;
  borderColor: string;
  bgGradientFrom: string;
  bgGradientTo: string;
  accessories: string[];
  createdAt: string;
}

export interface CustomSkinInput {
  /** @minLength 1 */
  name: string;
  glowColor: string;
  glowEnabled: boolean;
  borderColor: string;
  bgGradientFrom: string;
  bgGradientTo: string;
  accessories: string[];
}

export interface ResidentRecord {
  userId: number;
  robloxUsername: string;
  displayName: string;
  description: string;
}

export interface PinpointRecord {
  id: number;
  mapId: number;
  type: string;
  name: string;
  description: string;
  imageData: string;
  xPercent: number;
  yPercent: number;
  createdAt: string;
  residents: ResidentRecord[];
}

export interface MapRecord {
  id: number;
  name: string;
  imageData: string;
  createdAt: string;
  pinpoints: PinpointRecord[];
}

export interface MapInput {
  /** @minLength 1 */
  name: string;
  imageData?: string;
  accessCode?: string;
}

export interface PinpointInput {
  type?: string;
  /** @minLength 1 */
  name: string;
  description?: string;
  imageData?: string;
  xPercent?: number;
  yPercent?: number;
  accessCode?: string;
}

export interface LiveHereInput {
  /** @maxLength 500 */
  description?: string;
}

export interface UserLocationRecord {
  id: number;
  userId: number;
  pinpointId: number;
  description: string;
  createdAt: string;
}

export interface BadgeInventoryRecord {
  id: number;
  userId: number;
  badgeName: string;
  grantedAt: string;
}

export interface GiveBadgeInput {
  userId: number;
  badgeName: string;
  accessCode?: string;
}