# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
...
  /badge-inventory:
    get:
      operationId: listBadgeInventory
      tags: [badgeInventory]
      summary: Get current user badge inventory
      responses:
        "200":
          description: Badge inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BadgeInventoryRecord"
    post:
      operationId: giveBadge
      tags: [badgeInventory]
      summary: Give a badge to a user (moderator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiveBadgeInput"
      responses:
        "201":
          description: Badge given
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeInventoryRecord"
  /badge-inventory/user/{userId}:
    get:
      operationId: listBadgeInventoryForUser
      tags: [badgeInventory]
      summary: Get badge inventory for a specific user (mod panel)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Badge inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BadgeInventoryRecord"
  /badge-inventory/{id}:
    delete:
      operationId: removeBadge
      tags: [badgeInventory]
      summary: Remove a badge from inventory (moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AuthCredentials:
      type: object
      required: [robloxUsername, password]
      properties:
        robloxUsername:
          type: string
          minLength: 2
          maxLength: 30
        password:
          type: string
          minLength: 6
    UserRecord:
      type: object
      required: [id, robloxUsername, status]
      properties:
        id:
          type: integer
        robloxUsername:
          type: string
        status:
          type: string
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        status:
          type: string
          maxLength: 60
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserRecord"
...
    BadgeInventoryRecord:
      type: object
      required: [id, userId, badgeName, grantedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        badgeName:
          type: string
        grantedAt:
          type: string
    GiveBadgeInput:
      type: object
      required: [userId, badgeName]
      properties:
        userId:
          type: integer
        badgeName:
          type: string
        accessCode:
          type: string